requirements.txt
absl-py==0.9.0
asgiref==3.2.10
astor==0.8.0
astroid==2.3.3
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.1.5
blinker==1.4
brotlipy==0.7.0
cachetools==4.1.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==2.9.2
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.8
django-bootstrap4==2.2.0
entrypoints==0.3
environment-kernels==1.1.1
flickrapi==2.4.0
gast==0.2.2
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.10.0
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
importlib-metadata @ file:///C:/ci/importlib-metadata_1593446525189/work
ipykernel==5.3.0
ipython @ file:///C:/ci/ipython_1593433781269/work
ipython-genutils==0.2.0
isort==4.3.21
jedi @ file:///C:/ci/jedi_1592833832328/work
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lazy-object-proxy==1.5.0
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.1.0
packaging==20.4
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.3
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.4.4
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyreadline==2.1
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.1
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.0
scikit-learn==0.23.1
scipy @ file:///C:/ci/scipy_1592916958183/work
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.9
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==0.16.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
zipp==3.1.0
